+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 3.4.3
 ----------------
 - Fix #198 Add error if student account inactive (today < Attendance start date) in index.php
+- Optimization Remove $schools_RET & $calendars_RET ID index in Widgets.fnc.php
 
 Changes in 3.4.2
 ----------------